fix(mfa): move NextResponse-dependent buildEnrollOptions to server-only module
The client barrel (client/mfa/index.ts) imported buildEnrollOptions from
mfa-transform-utils.ts, which transitively pulled in next/server via a
top-level NextResponse import. This broke Jest jsdom tests with
ReferenceError: Request is not defined.

Move buildEnrollOptions (the only function using NextResponse) into a new
server-only file src/utils/mfa-server-utils.ts. The four client-safe
helpers (camelizeAuthenticator, camelizeChallengeResponse,
buildEnrollmentResponse, normalizeEnrollOptions) remain in
mfa-transform-utils.ts without any server-side imports.

Update src/server/auth-client.ts to import buildEnrollOptions from the
new module.

Fixes #2639
